ingestion/snowflake_pipeline/ingest_service/requirements.txt
snowflake-connector-python==3.12.0
pandas==2.2.2
pinecone-client==5.0.1
google-cloud-logging==3.11.2
google-cloud-aiplatform==1.64.0
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
llama-index-core==0.11.2
llama-index-readers-file==0.2.0
llama-index-vector-stores-pinecone==0.2.1
llama-index-embeddings-gemini==0.2.0
html2text==2024.2.26
transformers==4.44.2
